+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # HEAD
 
+-   Add `--license` command-line switch and include manpage in releases ([PR#1](https://github.com/ziotom78/audiotagedit/pull/1))
+
 # 0.1.0
 
 -   First release